updates
New two apis endpoint, uploading image on song and artist , on error show default image

+++ fontend/src/pages/Admin/pages/EditGenre.jsx
@@ -18,9 +18,15 @@ const EditGenre = () => {
   }, [reload]);
   const handleDelete = async (id) => {
     // genre/delete/4
-    const response = await deleteGenre(`/genre/delete/${id}`);
-    setReload(!reload);
-    alert(response);
+    const ok =  confirm("Are you sure you want to delete?")
+    if(ok){
+      const response = await deleteGenre(`/genre/delete/${id}`);
+      setReload(!reload);
+      alert("Deleted Genre");
+    }else{
+      return;
+    }
+  
   }
 // console.log(reload);
   return (
@@ -29,24 +35,24 @@ const EditGenre = () => {
         <p className="text-3xl font-bold text-center">Genre List</p>
       </div>
       <div className="mt-16 ml-10 ">
-        <div className="flex gap-10 flex-col w-full">
-          <div className="w-full relative"><button className="transition ease-in hover:scale-110 absolute bg-slate-600 px-5 rounded-2xl py-1 right-0"> <Link to="/admin/addGenre">Add</Link></button></div>
-          <table className="border-collapse  border-spacing-2 border w-full table-auto">
+        <div className="flex flex-col w-full gap-10">
+          <div className="relative w-full"><button className="absolute right-0 px-5 py-1 transition ease-in hover:scale-110 bg-slate-600 rounded-2xl"> <Link to="/admin/addGenre">Add</Link></button></div>
+          <table className="w-full border border-collapse table-auto border-spacing-2">
             <thead>
               <tr>
-                <th className="border border-gray-200 px-4 py-2">#</th>
-                <th className="border border-gray-200 px-4 py-2">Name</th>
-                <th className="border border-gray-200 px-4 py-2">Description</th>
-                <th className="border border-gray-200 px-4 py-2">Action</th>
+                <th className="px-4 py-2 border border-gray-200">#</th>
+                <th className="px-4 py-2 border border-gray-200">Name</th>
+                <th className="px-4 py-2 border border-gray-200">Description</th>
+                <th className="px-4 py-2 border border-gray-200">Action</th>
               </tr>
             </thead>
             <tbody>
               {genre.map((g,i) => (
                 <tr key={i}>
-                  <td className="border text-center px-4 py-2">{(i+1)}</td>
-                  <td className="border text-center px-4 py-2">{g.name}</td>
-                  <td className="border text-center px-4 py-2">{g.bio}</td>
-                  <td className="border flex justify-around px-4 py-2">
+                  <td className="px-4 py-2 text-center border">{(i+1)}</td>
+                  <td className="px-4 py-2 text-center border">{g.name}</td>
+                  <td className="px-4 py-2 text-center border">{g.bio}</td>
+                  <td className="flex justify-around px-4 py-2 border">
                     {" "}
                     <button className="text-xl transition-all ease-in hover:scale-125"><CiEdit /></button>
                     <button onClick={e=>handleDelete(g.id)}  className="text-xl transition-all ease-in hover:scale-125"><MdOutlineDelete /></button>
